fix: refactored constants
