--- src/register/RegistrationFields/CountryField/constants.js
@@ -1,2 +1,0 @@
-export const COUNTRY_CODE_KEY = 'code';
-export const COUNTRY_DISPLAY_KEY = 'name';
