+++ src/register/RegistrationFields/NameField/validator.js
@@ -1,5 +1,7 @@
-import { urlRegex } from './constants';
 import messages from '../../messages';
+
+export const INVALID_NAME_REGEX = /[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_\+.~#?&//=]*)?/gi; // eslint-disable-line no-useless-escape
+export const urlRegex = new RegExp(INVALID_NAME_REGEX);
 
 const validateName = (value, formatMessage) => {
   let fieldError;
