--- src/register/RegistrationFields/UsernameField/constants.js
@@ -1,3 +1,0 @@
-const VALID_USERNAME_REGEX = /^[a-zA-Z0-9_-]*$/i;
-
-export default VALID_USERNAME_REGEX;
